Res-contact/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rescontact"
version = "0.1.0"
description = "Protein contact map pipeline (ESM + FastAPI) with lightweight monitoring"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Ryoji Takahashi" }]

# Keep Torch out (platform-specific). Install Torch separately.
dependencies = [
  "fastapi>=0.111,<0.116",
  "uvicorn[standard]>=0.30,<0.33",
  "pydantic>=2.7,<3",
  "PyYAML>=6.0,<7",
  "numpy>=1.26,<3",
  "pandas>=2.2,<3",
  "scikit-learn>=1.4,<1.6",
  "biopython>=1.83,<2",
  "matplotlib>=3.8,<4",
  "transformers>=4.44,<5",
  "tokenizers>=0.19,<0.21",
  "regex>=2024.5.15",
  "tqdm>=4.66,<5",
  "optuna>=3.6,<4"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["rescontact*"]

[tool.setuptools.package-data]
"rescontact" = ["**/*.yaml", "**/*.yml"]

